SystemUI: fix recents task view height calculation
Initialize search bar height early enough for layouts to catch it and
not cause jumpiness in certain situations. For instance on a device with
a soft navigation bar, hitting recents from any application, right after
the animation would finish, it would relize its layout was incorrect and
relayout the thumbnails causing a jump.

Change-Id: Ic3d42a7efa7e66218ad9c8e364d90c981acc1ba6
Signed-off-by: Roman Birg <roman@cyngn.com>